Refactored entities so recipe is used to avoud API calls, UserRecipes are copies that user can edit/delete freely

+++ src/main/java/com/cakeplanner/cake_planner/Controller/ShoppingController.java
@@ -21,7 +21,7 @@ public class ShoppingController {
     }
 
     @PostMapping("/shopping/custom_list")
-    public String showUpdateForm(@RequestParam(value = "selectedCakeIds", required = false) List<Integer> selectedCakeIds,
+    public String showUpdateForm(@RequestParam(value = "selectedCakeIds", required = false) List<Long> selectedCakeIds,
                                  RedirectAttributes redirectAttributes,
                                  @ModelAttribute("user") User user) {
         if(selectedCakeIds.isEmpty()){
