+++ src/main/java/com/cakeplanner/cake_planner/Model/Repositories/ShoppingListRepository.java
@@ -3,6 +3,6 @@ package com.cakeplanner.cake_planner.Model.Repositories;
 import com.cakeplanner.cake_planner.Model.Entities.ShoppingList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShoppingListRepository extends JpaRepository<ShoppingList, Integer> {
+public interface ShoppingListRepository extends JpaRepository<ShoppingList, Long> {
 
 }
