+++ src/main/java/com/cakeplanner/cake_planner/Model/Repositories/UserRepository.java
@@ -3,8 +3,8 @@ package com.cakeplanner.cake_planner.Model.Repositories;
 import com.cakeplanner.cake_planner.Model.Entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     User findByEmail(String email);
 
-    User findByUserId(Integer id);
+    User findByUserId(Long id);
 }
