+++ src/main/java/com/cakeplanner/cake_planner/Model/Services/UserService.java
@@ -20,7 +20,7 @@ public class UserService {
         return user;
     }
 
-    public User findByUserId(Integer userId) {
+    public User findByUserId(Long userId) {
         User user = userRepository.findByUserId(userId);
         return user;
     }
